+++ helpers/pipeline_mode.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class PipelineMode(Enum):
+    NEW_PATRONS = 1
+    UPDATED_PATRONS = 2
+    DELETED_PATRONS = 3
+
+    def __str__(self):
+        return self.name.lower().rstrip('_patrons')